refactor: rework events handling logic
